weather-widget/requirements.txt
astroid==2.5.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
pycryptodome==3.10.1
pylint==2.7.2
PyYAML==5.4.1
requests==2.25.1
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
